Moved all standalone tests to test folder
