+++ test/test-duplicate-nested-expressions.js
@@ -26,11 +26,11 @@
 
 var assert = require('assert');
 var extend = require('extend');
-var v8debugapi = require('../../src/agent/v8debugapi.js');
+var v8debugapi = require('../src/agent/v8debugapi.js');
 var common = require('@google-cloud/common');
-var defaultConfig = require('../../src/agent/config.js');
-var SourceMapper = require('../../src/agent/sourcemapper.js');
-var scanner = require('../../src/agent/scanner.js');
+var defaultConfig = require('../src/agent/config.js');
+var SourceMapper = require('../src/agent/sourcemapper.js');
+var scanner = require('../src/agent/scanner.js');
 var path = require('path');
 var semver = require('semver');
 
@@ -44,7 +44,7 @@ function stateIsClean(api) {
 
 describe('v8debugapi', function() {
   var config = extend({}, defaultConfig, {
-    workingDirectory: path.join(process.cwd(), 'test', 'standalone')
+    workingDirectory: path.join(process.cwd(), 'test')
   });
   var logger = common.logger({ logLevel: config.logLevel });
   var api = null;
