+++ test/test-max-data-size.js
@@ -24,10 +24,10 @@ process.env.GCLOUD_DIAGNOSTICS_CONFIG = 'test/fixtures/test-config.js';
 var assert = require('assert');
 var extend = require('extend');
 var common = require('@google-cloud/common');
-var v8debugapi = require('../../src/agent/v8debugapi.js');
-var SourceMapper = require('../../src/agent/sourcemapper.js');
-var scanner = require('../../src/agent/scanner.js');
-var defaultConfig = require('../../src/agent/config.js');
+var v8debugapi = require('../src/agent/v8debugapi.js');
+var SourceMapper = require('../src/agent/sourcemapper.js');
+var scanner = require('../src/agent/scanner.js');
+var defaultConfig = require('../src/agent/config.js');
 var api;
 
 var breakpointInFoo = {
