+++ test/test-this-context.js
@@ -1,12 +1,12 @@
-/*1* KEEP THIS CODE AT THE TOP TO AVOID LINE NUMBER CHANGES */ /* jshint shadow:true */
+/*1* KEEP THIS CODE AT THE TOP TO AVOID LINE NUMBER CHANGES */
 /*2*/'use strict';
-/*3*/function foo() {
-/*4*/ try {
-/*5*/   throw new Error('A test');
-/*6*/ } catch (e) {
-/*7*/   var e = 2;
-/*8*/   return e;
-/*9*/ }
+/*3*/function foo(b) {/* jshint validthis: true */
+/*4*/ this.a = 10;
+/*5*/ this.a += b;
+/*6*/ return this;
+/*7*/}
+/*8*/function bar(j) {
+/*9*/ return j;
 /*10*/}
 /**
  * Copyright 2015 Google Inc. All Rights Reserved.
@@ -26,11 +26,11 @@
 
 var assert = require('assert');
 var extend = require('extend');
-var v8debugapi = require('../../src/agent/v8debugapi.js');
+var v8debugapi = require('../src/agent/v8debugapi.js');
 var common = require('@google-cloud/common');
-var defaultConfig = require('../../src/agent/config.js');
-var SourceMapper = require('../../src/agent/sourcemapper.js');
-var scanner = require('../../src/agent/scanner.js');
+var defaultConfig = require('../src/agent/config.js');
+var SourceMapper = require('../src/agent/sourcemapper.js');
+var scanner = require('../src/agent/scanner.js');
 var path = require('path');
 var semver = require('semver');
 
@@ -44,7 +44,7 @@ function stateIsClean(api) {
 
 describe('v8debugapi', function() {
   var config = extend({}, defaultConfig, {
-    workingDirectory: path.join(process.cwd(), 'test', 'standalone')
+    workingDirectory: path.join(process.cwd(), 'test')
   });
   var logger = common.logger({ logLevel: config.logLevel });
   var api = null;
@@ -71,11 +71,12 @@ describe('v8debugapi', function() {
     }
   });
   afterEach(function() { assert(stateIsClean(api)); });
-  it('Should read e as the caught error', function(done) {
-    var brk = {
-      id: 'fake-id-123',
-      location: { path: 'test-try-catch.js', line: 7 }
-    };
+  it('Should be able to read the argument and the context', function(done) {
+      var brk = {
+        id: 'fake-id-123',
+        location: { path: 'test-this-context.js', line: 5 }
+      };
+      var ctxMembers;
     api.set(brk, function(err) {
       assert.ifError(err);
       api.wait(brk, function(err) {
@@ -83,31 +84,36 @@ describe('v8debugapi', function() {
         var frame = brk.stackFrames[0];
         var args = frame.arguments;
         var locals = frame.locals;
-        assert.equal(locals.length, 1, 'There should be one local');
+        ctxMembers = brk.variableTable.slice(brk.variableTable.length-1)[0]
+          .members;
+        assert.deepEqual(ctxMembers.length, 1, 
+          'There should be one member in the context variable value');
+        assert.deepEqual(ctxMembers[0], {name: 'a', value: '10'});
         if (semver.satisfies(process.version, '<1.6')) {
-          // Try/Catch scope-walking does not work on 0.12
-          assert.deepEqual(
-            locals[0],
-            {name: 'e', value: 'undefined'}
-          );
+            assert.equal(args.length, 1, 'There should be one argument');
+            assert.equal(locals.length, 1, 'There should be one local');
+            assert.deepEqual(
+              args[0],
+              {name: 'b', value: '1'}
+   	   	    );
+            assert.deepEqual(locals[0].name, 'context');
         } else {
           assert.equal(args.length, 0, 'There should be zero arguments');
-          var e = locals[0];
-          assert(e.name === 'e');
-          assert(Number.isInteger(e.varTableIndex));
+          assert.equal(locals.length, 2, 'There should be two locals');
+          assert.deepEqual(locals[0], {name: 'b', value: '1'});
+          assert.deepEqual(locals[1].name, 'context');
         }
-        assert.equal(args.length, 0, 'There should be zero arguments');     
         api.clear(brk);
         done();
       });
-      process.nextTick(foo.bind(null, 'test'));
+      process.nextTick(foo.bind({}, 1));
     });
   });
-  it('Should read e as the local error', function(done) {
-    var brk = {
-      id: 'fake-id-123',
-      location: { path: 'test-try-catch.js', line: 8 }
-    };
+  it('Should be able to read the argument and deny the context', function(done) {
+      var brk = {
+        id: 'fake-id-123',
+        location: { path: 'test-this-context.js', line: 9 }
+      };
     api.set(brk, function(err) {
       assert.ifError(err);
       api.wait(brk, function(err) {
@@ -116,24 +122,21 @@ describe('v8debugapi', function() {
         var args = frame.arguments;
         var locals = frame.locals;
         if (semver.satisfies(process.version, '<1.6')) {
-          // Try/Catch scope-walking does not work on 0.12
-          assert.deepEqual(
-            locals[0],
-            {name: 'e', value: 'undefined'}
-          );
+            assert.equal(args.length, 1, 'There should be one argument');
+            assert.equal(locals.length, 0);
+             assert.deepEqual(
+              args[0],
+              {name: 'j', value: '1'}
+   	   	    );
         } else {
           assert.equal(args.length, 0, 'There should be zero arguments');
           assert.equal(locals.length, 1, 'There should be one local');
-          assert.deepEqual(
-            locals[0],
-            {name: 'e', value: '2'}
-          );
-          assert.equal(args.length, 0, 'There should be zero arguments');
+          assert.deepEqual(locals[0], {name: 'j', value: '1'});
         }
         api.clear(brk);
         done();
       });
-      process.nextTick(foo.bind(null, 'test'));
+      process.nextTick(bar.bind(null, 1));
     });
   });
 });
